+++ hardening/Root Locker/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.7 - 2024-08-15
+
+### Changed
+
+- Improved error handling.
+- Modify syntax and documentation.
+- Utilizes `usermod -L` to lock the root account.
+- Rename script to `root-locker.bash`.
+
 ## v1.0.6 - 2024-04-13
 
 ### Changed